requirements.txt
opencv-python>=4.5.5
pytesseract>=0.3.10
numpy>=1.21.0
python-escpos>=3.0.0
